* modify console application * little refactoring

+++ NET.W.2018.Соколовский.08/NET.W.2018.Соколовский.08.BusinessLogic/Services/BankAccountService.cs
@@ -1,4 +1,6 @@
-﻿using NET.W._2018.Соколовский._08.Common.Interfaces.Repositories;
+﻿using System;
+using System.Collections.Generic;
+using NET.W._2018.Соколовский._08.Common.Interfaces.Repositories;
 using NET.W._2018.Соколовский._08.Common.Interfaces.Services;
 using NET.W._2018.Соколовский._08.Common.Models;
 
@@ -14,12 +16,22 @@ namespace NET.W._2018.Соколовский._08.BusinessLogic.Services
         }
         public BankAccount AddAccount(BankAccount bankAccount)
         {
+            if (ReferenceEquals(bankAccount, null))
+            {
+                throw new ArgumentNullException(nameof(bankAccount));
+            }
+
             return this._storage.Add(bankAccount);
         }
 
         public BankAccount CloseAnAccount(int accountId)
         {
             var updatedBankAccount = this._storage.Get(accountId);
+            if (ReferenceEquals(updatedBankAccount, null))
+            {
+                throw new ArgumentException(nameof(accountId));
+            }
+
             updatedBankAccount.IsClosed = true;
             return this._storage.Update(updatedBankAccount);
         }
@@ -33,14 +45,28 @@ namespace NET.W._2018.Соколовский._08.BusinessLogic.Services
 
         public BankAccount RemoveAccount(BankAccount bankAccount)
         {
+            if (ReferenceEquals(bankAccount, null))
+            {
+                throw new ArgumentNullException(nameof(bankAccount));
+            }
+
             return this._storage.Delete(bankAccount);
         }
 
         public BankAccount TopUpInAnAccount(decimal amount, int accountId)
         {
             var updatedBankAccount = this._storage.Get(accountId);
+            if (ReferenceEquals(updatedBankAccount, null))
+            {
+                throw new ArgumentException(nameof(accountId));
+            }
             updatedBankAccount.Ammount = updatedBankAccount.Ammount + updatedBankAccount.Bonus + amount;
             return this._storage.Update(updatedBankAccount);
+        }
+
+        public IEnumerable<BankAccount> GetAllBankAccounts()
+        {
+            return this._storage.GetAll();
         }
     }
 }
